refactor(arch): move encoding.py to bytecode/ with re-export shim (stage 1 of #569)
First stage of the multi-stage package merge tracked in #569. Moves
the canonical home of LocationAddress, LaneAddress, ZoneAddress,
SiteLaneAddress, WordLaneAddress, Direction, MoveType from
bloqade.lanes.layout.encoding to bloqade.lanes.bytecode.encoding —
foundation types live alongside the Rust _native bindings they wrap.

Old import path bloqade.lanes.layout.encoding remains as a thin
re-export shim, so every existing caller across bloqade.lanes,
bloqade.gemini, demos, and tests compiles unchanged. The shim is
removed in the final cleanup stage (Stage 9 of the plan) once all
in-flight branches have rebased onto the canonical path.

Notable non-changes:
- bloqade/lanes/bytecode/__init__.py is left alone. It currently
  re-exports LocationAddress / LaneAddress / ZoneAddress / Direction /
  MoveType directly from _native (the Rust raw types). Switching
  the source to .encoding would change the class object callers
  receive (raw Rust class -> Python wrapper) — a real semantic
  break. That switch belongs in Stage 9 (or as a separate cleanup),
  not this PR.

Verification: 1099 pytest pass, pyright + ruff clean, six representative
non-display demos run end-to-end (the cudaq demo fails for a
pre-existing environment reason; the matplotlib-display demos are
out of scope for headless smoke testing).

Refs #569.
